Add typedefs from simple-hl7 PR
I found a pending PR for simple-hl7 which had 90% working typedefs
Awesome work! I needed to do some light massaging to get it to integrate,
but it seems to be okay.

I've probably broken a heap, so we'll need to review this now...
Reference PR: https://github.com/hitgeek/simple-hl7/pull/69/files